consumer/requirements.txt
kafka-python==2.0.2
PyYAML==6.0.1
psycopg2-binary==2.9.9
